Make the input hierarchical.
