+++ src/ttsim/__init__.py
@@ -1,21 +1,11 @@
 from __future__ import annotations
 
-from ttsim.interface_dag import (
-    InputDataDfAndMapper,
-    InputDataDfWithNestedColumns,
-    InputDataFlat,
-    InputDataQName,
-    InputDataTree,
-    main,
-)
+from ttsim.argument_templates import input_data
+from ttsim.interface_dag import main
 from ttsim.interface_dag_elements.shared import merge_trees
 
 __all__ = [
-    "InputDataDfAndMapper",
-    "InputDataDfWithNestedColumns",
-    "InputDataFlat",
-    "InputDataQName",
-    "InputDataTree",
+    "input_data",
     "main",
     "merge_trees",
 ]
