+++ src/ttsim/interface_dag.py
@@ -2,15 +2,15 @@ from __future__ import annotations
 
 import inspect
 import re
-from dataclasses import asdict, dataclass
+from dataclasses import asdict
 from pathlib import Path
 from typing import TYPE_CHECKING, Any, Literal
 
 import dags
 import dags.tree as dt
 import optree
-import pandas as pd
 
+from ttsim.argument_templates import input_data
 from ttsim.interface_dag_elements import _InterfaceDAGElements
 from ttsim.interface_dag_elements.fail_if import (
     format_errors_and_warnings,
@@ -28,72 +28,19 @@ if TYPE_CHECKING:
     import datetime
 
     from ttsim.interface_dag_elements.typing import (
-        FlatData,
-        NestedData,
         NestedTargetDict,
-        QNameData,
         QNameStrings,
         UnorderedQNames,
     )
 
 
-def _camel_to_snake(name: str) -> str:
-    # Insert underscore before each uppercase letter preceded by a lowercase letter or digit
-    s1 = re.sub("(.)([A-Z][a-z]+)", r"\1_\2", name)
-    # Insert underscore before uppercase letters preceded by lowercase letters or digits
-    s2 = re.sub("([a-z0-9])([A-Z])", r"\1_\2", s1)
-    # Convert the entire string to lowercase
-    return s2.lower()
-
-
-@dataclass(frozen=True)
-class InputDataABC:
-    def to_dict(self) -> dict[str, Any]:
-        name = _camel_to_snake(self.__class__.__name__[9:])
-        if len(self.__dict__) == 1:
-            return {name: self.data}  # type: ignore[attr-defined]
-        return {name: {k: v for k, v in self.__dict__.items()}}
-
-
-@dataclass(frozen=True)
-class InputDataDfAndMapper(InputDataABC):
-    df: pd.DataFrame
-    """A dataframe with arbitrary columns."""
-    mapper: dict[str, Any]
-    """A nested dictionary mapping expected inputs to column names in df."""
-
-
-@dataclass(frozen=True)
-class InputDataDfWithNestedColumns(InputDataABC):
-    data: pd.DataFrame
-    """A df with a MultiIndex in the column dimension, elements correspond to expected tree paths."""  # noqa: E501
-
-
-@dataclass(frozen=True)
-class InputDataTree(InputDataABC):
-    data: NestedData
-    """A nested dictionary mapping expected input names to vectors of data."""
-
-
-@dataclass(frozen=True)
-class InputDataFlat(InputDataABC):
-    data: FlatData
-    """A dictionary mapping tree paths to vectors of data."""
-
-
-@dataclass(frozen=True)
-class InputDataQName(InputDataABC):
-    data: QNameData
-    """A dictionary mapping qualified names to vectors of data."""
-
-
 def main(
     date_str: str | None = None,
     output_names: NestedTargetDict | QNameStrings | None = None,
-    input_data: InputDataDfAndMapper
-    | InputDataDfWithNestedColumns
-    | InputDataFlat
-    | InputDataQName
+    input_data: input_data.DfAndMapper
+    | input_data.DfWithNestedColumns
+    | input_data.Flat
+    | input_data.QName
     | None = None,
     targets: dict[str, Any] | None = None,
     backend: Literal["numpy", "jax"] | None = None,
@@ -169,7 +116,7 @@ def _harmonize_inputs(inputs: dict[str, Any]) -> dict[str, Any]:
         asdict(_InterfaceDAGElements()),  # type: ignore[arg-type]
         none_is_leaf=True,
     )[:2]
-    if "input_data" in inputs and isinstance(inputs["input_data"], InputDataABC):
+    if "input_data" in inputs and isinstance(inputs["input_data"], input_data.ABC):
         inputs["input_data"] = inputs["input_data"].to_dict()
     for acc, val in zip(accs, vals, strict=False):
         qname = dt.qname_from_tree_path(acc.path)
