+++ tests/ttsim/test_end_to_end.py
@@ -4,8 +4,7 @@ from typing import Literal
 import pandas as pd
 import pytest
 
-from ttsim import main
-from ttsim.interface_dag import InputDataDfAndMapper, InputDataDfWithNestedColumns
+from ttsim import input_data, main
 
 DF_WITH_NESTED_COLUMNS = pd.DataFrame(
     {
@@ -79,19 +78,19 @@ EXPECTED_RESULTS = pd.DataFrame(
 
 
 @pytest.mark.parametrize(
-    "input_data",
+    "input_data_arg",
     [
         # Correct way to do it
-        InputDataDfAndMapper(df=DF_FOR_MAPPER, mapper=INPUT_DF_MAPPER),
-        InputDataDfWithNestedColumns(data=DF_WITH_NESTED_COLUMNS),
+        input_data.DfAndMapper(df=DF_FOR_MAPPER, mapper=INPUT_DF_MAPPER),
+        input_data.DfWithNestedColumns(data=DF_WITH_NESTED_COLUMNS),
         # May or may not continue to work.
         {"df_and_mapper": {"df": DF_FOR_MAPPER, "mapper": INPUT_DF_MAPPER}},
         {"df_with_nested_columns": DF_WITH_NESTED_COLUMNS},
     ],
 )
-def test_end_to_end(input_data, backend: Literal["numpy", "jax"]):
+def test_end_to_end(input_data_arg, backend: Literal["numpy", "jax"]):
     result = main(
-        input_data=input_data,
+        input_data=input_data_arg,
         targets={"tree": TARGETS_TREE},
         date_str="2025-01-01",
         rounding=False,
